backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==4.6.0
djongo==1.3.6
python-decouple==3.8
PyJWT==2.8.0
django-otp==1.2.3
python-dotenv==1.0.0
pillow>=12.0.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
qrcode[pil]==7.4.2
requests==2.32.3

